fix state root
